// resources/js/admin/game-maps/components/side-peeks/enums/game-map-side-peek-messages.ts
export enum GameMapSidePeekMessages {
  NoEntitiesAtCoordinate = 'No Locations or Npcs occupy this coordinate.',
  NoLocationsToMove = 'There are no Locations on this Game Map to move.',
  NoNpcsToMove = 'There are no Npcs on this Game Map to move.',
  SearchLocations = 'Search Locations',
  SearchNpcs = 'Search Npcs',
  LocationMoved = 'Location moved.',
  NpcMoved = 'Npc moved.',
  LocationSaved = 'Location saved.',
  NpcSaved = 'Npc saved.',
  GameMapSaved = 'Game Map saved.',
  SelectDestinationInstruction = 'Select a valid coordinate on the map to choose a destination.',
}