// resources/js/admin/game-maps/components/side-peeks/enums/game-map-side-peek-view.ts
export enum GameMapSidePeekView {
  Details = 'details',
  CreateLocation = 'create-location',
  CreateNpc = 'create-npc',
  SelectLocationToMove = 'select-location-to-move',
  SelectNpcToMove = 'select-npc-to-move',
  ConfirmMoveLocation = 'confirm-move-location',
  ConfirmMoveNpc = 'confirm-move-npc',
  EditLocation = 'edit-location',
  EditNpc = 'edit-npc',
}